doc: Separate code snippet into file
