+++ apps/insight-viewer-demo/src/containers/Basic/Code.ts
@@ -1,0 +1,14 @@
+export const DICOM_CODE = `\
+import Viewer from '@lunit/insight-viewer'
+
+export default function() {
+  return <Viewer.Dicom imageId={IMAGE_ID} />
+}
+`
+export const WEB_CODE = `\
+import Viewer from '@lunit/insight-viewer'
+
+export default function() {
+  return <Viewer.Web imageId={WEB_URL} />
+}
+`
