+++ apps/insight-viewer-demo/src/containers/Error/Custom.tsx
@@ -1,18 +1,7 @@
 import { Box } from '@chakra-ui/react'
 import Viewer, { ViewerError } from '@lunit/insight-viewer'
 import CodeBlock from '../../components/CodeBlock'
-
-const code = `\
-import Viewer, { ViewerError } from '@lunit/insight-viewer'
-
-function customError(e: ViewerError): void {
-  alert(e.message + e.status)
-}
-
-export default function() {
-  return <Viewer.Dicom imageId={IMAGE_ID} onError={customError} />
-}
-`
+import { CUSTOM_CODE } from './Code'
 
 const IMAGE_ID =
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT0000100.dcm'
@@ -29,7 +18,7 @@ export default function Custom(): JSX.Element {
         <Viewer.Dicom imageId={IMAGE_ID} onError={customError} />
       </Box>
       <Box w={800}>
-        <CodeBlock code={code} />
+        <CodeBlock code={CUSTOM_CODE} />
       </Box>
     </>
   )
