+++ apps/insight-viewer-demo/src/containers/HttpHeader/Code.ts
@@ -1,0 +1,16 @@
+export const CODE = `\
+import Viewer from '@lunit/insight-viewer'
+
+const requestInterceptor = (request: Request) => {
+  request.headers.set('Authorization', 'Bearer blahblah')
+}
+
+export default function() {
+  return (
+    <Viewer.Dicom 
+      imageId={IMAGE_ID} 
+      requestInterceptor={requestInterceptor} 
+    />
+  )
+}
+`
