+++ apps/insight-viewer-demo/src/containers/Interaction/Base.tsx
@@ -8,55 +8,10 @@ import CodeBlock from '../../components/CodeBlock'
 import Control from './Control'
 import OverlayLayer from '../../components/OverlayLayer'
 import CustomProgress from '../../components/CustomProgress'
+import { BASE_CODE } from './Code'
 
 const IMAGE_ID =
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000011.dcm'
-
-const Code = `\
-  import Viewer, { useInteraction, useViewport, Interaction } from '@lunit/insight-viewer'
-
-  export default function App() {
-    const { interaction, setInteraction } = useInteraction()
-    const { viewport, setViewport } = useViewport()
-
-    function handlePrimaryDrag(e: React.ChangeEvent<HTMLInputElement>) {
-      const v = e.target.value
-
-      setInteraction((prev: Interaction) => ({
-        ...prev,
-        primaryDrag: v === 'none' ? undefined : v,
-      }))
-    }
-
-    function handleSecondaryDrag(e: React.ChangeEvent<HTMLInputElement>) {
-      const v = e.target.value
-
-      setInteraction((prev: Interaction) => ({
-        ...prev,
-        secondaryDrag: v === 'none' ? undefined : v,
-      }))
-    }
-
-    return (
-      <>
-        <input type="radio" value="none" onChange={handlePrimaryDrag} />
-        <input type="radio" value="pan" onChange={handlePrimaryDrag} />
-        <input type="radio" value="adjust" onChange={handlePrimaryDrag} />
-        <input type="radio" value="none" onChange={handleSecondaryDrag} />
-        <input type="radio" value="pan" onChange={handleSecondaryDrag} />
-        <input type="radio" value="adjust" onChange={handleSecondaryDrag} />
-        <Viewer.Dicom
-          imageId={IMAGE_ID}
-          interaction={interaction}
-          viewport={viewport}
-          onViewportChange={setViewport}
-        >
-          <OverlayLayer viewport={viewport} />
-        </Viewer.Dicom>
-      </>
-    )
-  }
-  `
 
 export default function App(): JSX.Element {
   const { interaction, setInteraction } = useInteraction()
@@ -86,7 +41,7 @@ export default function App(): JSX.Element {
         </Viewer.Dicom>
       </Box>
       <Box w={900}>
-        <CodeBlock code={Code} />
+        <CodeBlock code={BASE_CODE} />
       </Box>
     </Box>
   )
