+++ apps/insight-viewer-demo/src/containers/Interaction/Custom.tsx
@@ -14,92 +14,10 @@ import Control from './Control'
 import ClickControl from './Control/Click'
 import OverlayLayer from '../../components/OverlayLayer'
 import CustomProgress from '../../components/CustomProgress'
+import { CUSTOM_CODE } from './Code'
 
 const IMAGE_ID =
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000011.dcm'
-
-const Code = `\
-  import Viewer, { useInteraction, Interaction, Drag } from '@lunit/insight-viewer'
-
-  export default function App() {
-    const { interaction, setInteraction } = useInteraction()
-    const { viewport, setViewport } = useViewport()
-
-    const customPan: Drag = ({ viewport, delta }) => {
-      console.log(
-        'pan',
-        viewport.translation.x,
-        viewport.translation.y,
-        delta.x,
-        delta.y
-      )
-      setViewport(prev => ({
-        ...prev,
-        x: prev.x + delta.x / prev.scale,
-        y: prev.y + delta.y / prev.scale,
-      }))
-    }
-
-    const customAdjust: Drag = ({ viewport, delta }) => {
-      console.log(
-        'adjust',
-        viewport.voi.windowWidth,
-        viewport.voi.windowCenter,
-        delta.x,
-        delta.y
-      )
-      setViewport(prev => ({
-        ...prev,
-        windowWidth: prev.windowWidth + delta.x / prev.scale,
-        windowCenter: prev.windowCenter + delta.y / prev.scale,
-      }))
-    }
-
-    function handleCustomPan(e: React.ChangeEvent<HTMLInputElement>) {
-      setInteraction((prev: Interaction) => ({
-        ...prev,
-        primaryDrag: e.target.value === 'none' ? undefined : customPan,
-      }))
-    }
-
-    function handleCustomAdjust(e: React.ChangeEvent<HTMLInputElement>) {
-      setInteraction((prev: Interaction) => ({
-        ...prev,
-        secondaryDrag: e.target.value === 'none' ? undefined : customAdjust,
-      }))
-    }
-
-    function handleClick(e) {
-      if (e.target.checked) {
-        setInteraction((prev: Interaction) => ({
-          ...prev,
-          primaryClick: // or secondaryClick
-            value === 'none'
-              ? undefined
-              : (offsetX, offsetY) => {
-                console.log(offsetX, offsetY)
-              },
-        }))
-      }
-    }
-
-    return (
-      <>
-        <input type="radio" value="pan" onChange={handleCustomPan} />
-        <input type="radio" value="adjust" onChange={handleCustomAdjust} />
-        <input type="checkbox" value="adjust" onChange={handleClick} />
-        <Viewer.Dicom
-          imageId={IMAGE_ID}
-          interaction={interaction}
-          viewport={viewport}
-          onViewportChange={setViewport}
-        >
-          <OverlayLayer viewport={viewport} />
-        </Viewer.Dicom>
-      </>
-    )
-  }
-  `
 
 export default function App(): JSX.Element {
   const [{ x, y }, setCoord] = useState({ x: 0, y: 0 })
@@ -203,7 +121,7 @@ export default function App(): JSX.Element {
         </Viewer.Dicom>
       </Box>
       <Box w={900}>
-        <CodeBlock code={Code} />
+        <CodeBlock code={CUSTOM_CODE} />
       </Box>
     </Box>
   )
