+++ apps/insight-viewer-demo/src/containers/MultiFrame/Base.tsx
@@ -3,6 +3,7 @@ import Viewer, { useMultiframe } from '@lunit/insight-viewer'
 import React from 'react'
 import CodeBlock from '../../components/CodeBlock'
 import CustomProgress from '../../components/CustomProgress'
+import { CODE } from './Code'
 
 const IMAGES = [
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000000.dcm',
@@ -17,48 +18,6 @@ const IMAGES = [
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000009.dcm',
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000010.dcm',
 ]
-
-const Code = `\
-import Viewer, { useMultiframe } from '@lunit/insight-viewer'
-
-const IMAGES = [
-  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000000.dcm',
-  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000001.dcm',
-  ...
-]
-
-export default function Viewer() {
-  const { image, frame, setFrame } = useMultiframe(
-    IMAGES, 
-    {
-      initialFrame,       // optional: initialValue or default 0
-      prefetch,           // optional: default true
-      onError,            // optional
-      requestInterceptor, // optional
-    } // optional
-  )
-
-  function changeFrame(e) {
-    setFrame(Number(e.target.value))
-  }
-
-  return (
-    <>
-      <input
-        type="range"
-        id="frame"
-        name="frame"
-        min="0"
-        max="10"
-        step="1"
-        defaultValue={0}
-        onChange={changeFrame}
-      />
-      <Viewer.Dicom imageId={image} /* or imageId={IMAGES[frame]} */ />
-    </>
-  )
-}
-`
 
 export default function Base(): JSX.Element {
   const { image, setFrame } = useMultiframe(IMAGES)
@@ -87,7 +46,7 @@ export default function Base(): JSX.Element {
       </Box>
       <Viewer.Dicom imageId={image} Progress={CustomProgress} />
       <Box w={900}>
-        <CodeBlock code={Code} />
+        <CodeBlock code={CODE} />
       </Box>
     </Box>
   )
