+++ apps/insight-viewer-demo/src/containers/Overlay/index.tsx
@@ -3,27 +3,10 @@ import Viewer, { useViewport } from '@lunit/insight-viewer'
 import CodeBlock from '../../components/CodeBlock'
 import OverlayLayer from '../../components/OverlayLayer'
 import CustomProgress from '../../components/CustomProgress'
+import { CODE } from './Code'
 
 const IMAGE_ID =
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000011.dcm'
-
-const Code = `\
-  import Viewer, { useViewport } from '@lunit/insight-viewer'
-
-  export default function Viewer() {
-    const { viewport, setViewport } = useViewport()
-
-    return (
-      <Viewer.Dicom 
-        imageId={IMAGE_ID} 
-        viewport={viewport}
-        onViewportChange={setViewport}
-      >
-        <OverlayLayer viewport={viewport} />
-      </Viewer.Dicom>
-    )
-  }
-  `
 
 function Overlay(): JSX.Element {
   const { viewport, setViewport } = useViewport()
@@ -41,7 +24,7 @@ function Overlay(): JSX.Element {
         </Viewer.Dicom>
       </Box>
       <Box w={900}>
-        <CodeBlock code={Code} />
+        <CodeBlock code={CODE} />
       </Box>
     </Box>
   )
