+++ apps/insight-viewer-demo/src/containers/Progress/Base.tsx
@@ -1,14 +1,7 @@
 import { Box } from '@chakra-ui/react'
 import Viewer from '@lunit/insight-viewer'
 import CodeBlock from '../../components/CodeBlock'
-
-const Code = `\
-import Viewer from '@lunit/insight-viewer'
-
-export default function Viewer() {
-  return <Viewer.Dicom imageId={IMAGE_ID} />
-}
-`
+import { BASE_CODE } from './Code'
 
 const IMAGE_ID =
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000011.dcm'
@@ -20,7 +13,7 @@ export default function Base(): JSX.Element {
         <Viewer.Dicom imageId={IMAGE_ID} />
       </Box>
       <Box w={700}>
-        <CodeBlock code={Code} />
+        <CodeBlock code={BASE_CODE} />
       </Box>
     </>
   )
