+++ apps/insight-viewer-demo/src/containers/Progress/Code.ts
@@ -1,0 +1,19 @@
+export const BASE_CODE = `\
+import Viewer from '@lunit/insight-viewer'
+
+export default function Viewer() {
+  return <Viewer.Dicom imageId={IMAGE_ID} />
+}
+`
+export const CUSTOM_CODE = `\
+import Viewer from '@lunit/insight-viewer'
+
+export default function Viewer() {
+  return (
+    <Viewer.Dicom 
+      imageId={IMAGE_ID} 
+      Progress={CustomProgress} 
+    />
+  )
+}
+`
