+++ apps/insight-viewer-demo/src/containers/Progress/Custom.tsx
@@ -2,19 +2,8 @@ import { Box } from '@chakra-ui/react'
 import Viewer from '@lunit/insight-viewer'
 import CodeBlock from '../../components/CodeBlock'
 import CustomProgress from '../../components/CustomProgress'
+import { CUSTOM_CODE } from './Code'
 
-const Code = `\
-import Viewer from '@lunit/insight-viewer'
-
-export default function Viewer() {
-  return (
-    <Viewer.Dicom 
-      imageId={IMAGE_ID} 
-      Progress={CustomProgress} 
-    />
-  )
-}
-`
 const IMAGE_ID =
   'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000012.dcm'
 
@@ -25,7 +14,7 @@ export default function Custom(): JSX.Element {
         <Viewer.Dicom imageId={IMAGE_ID} Progress={CustomProgress} />
       </Box>
       <Box w={700}>
-        <CodeBlock code={Code} />
+        <CodeBlock code={CUSTOM_CODE} />
       </Box>
     </>
   )
